Move some Render traits to Mods.Common
