+++ OpenRA.Mods.Common/Traits/Render/RenderNameTag.cs
@@ -14,7 +14,7 @@ using OpenRA.Graphics;
 using OpenRA.Mods.Common.Graphics;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Render
+namespace OpenRA.Mods.Common.Traits
 {
 	[Desc("Displays the player name above the unit")]
 	class RenderNameTagInfo : ITraitInfo
